--- a/julia/Related/Project.toml
+++ b/julia/Related/Project.toml
@@ -7,10 +7,11 @@ version = "0.1.0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
 
 [compat]
-julia = "1.9"
 JSON3 = "1"
 PrecompileTools = "1"
 StructTypes = "1"
+julia = "1.9"