Following the latest discussion in:
- https://github.com/jinyus/related_post_gen/issues/194#issuecomment-1777172218

@algunion @LilithHafner