@jinyus 2023-10-27
Main:
`Julia | 30.38 ms | 481.33 ms | 4.23 s`

This PR:
```
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 29 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
          Time (mean ± σ):     990.7 ms ±  12.9 ms    [User: 883.0 ms, System: 211.2 ms]
          Range (min … max):   973.7 ms … 1012.2 ms    10 runs
         
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 367 milliseconds
        Processing time (w/o IO): 389 milliseconds
        Processing time (w/o IO): 366 milliseconds
          Time (mean ± σ):      1.487 s ±  0.008 s    [User: 1.334 s, System: 0.255 s]
          Range (min … max):    1.481 s …  1.493 s    2 runs
         
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 3157 milliseconds
        Processing time (w/o IO): 3158 milliseconds
        Processing time (w/o IO): 3156 milliseconds
          Time (mean ± σ):      4.585 s ±  0.013 s    [User: 4.353 s, System: 0.327 s]
          Range (min … max):    4.576 s …  4.594 s    2 runs
         
```