@Moelf 2023-10-27
```
> math '30+389+3158'
3577
```

would rought put Julia somewhere around 5th? not bad I guess