@algunion 2023-10-27
> ```
> > math '30+389+3158'
> 3577
> ```
> 
> would roughly put Julia somewhere around 5th? not bad I guess

@Moelf , I am not sure this is optimal.

The _original_ StaticArrays implementation constantly achieved around 23ms for 5K posts (which is actually the value of first place right now - which is Zig).

37.15 ms is quite departing from that result (but I cannot say why - I just took a peek at the code).

I don't have too much time now - but if you have and are willing, you might try to reconcile the past [top result from here](https://github.com/algunion/related_post_gen/blob/rem-btime/julia/related.jl) with the current precompilation setup.