@Moelf 2023-10-28
I don't want to add the dynamic dispatch to different type length, because other implementations don't do that I believe