@Tortar 2023-10-28
Strangely it seems to me that the Julia version didn't even improve after merging this, the sum of the times is still nearly 5 seconds